dnsmadeeasy: re-organize the bins
cc/ @soniah
