+++ builtin/bins/provider-dnsmadeeasy/main.go
@@ -1,11 +1,12 @@
 package main
 
 import (
+	"github.com/hashicorp/terraform/builtin/providers/dnsmadeeasy"
 	"github.com/hashicorp/terraform/plugin"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
-		ProviderFunc: Provider,
+		ProviderFunc: dnsmadeeasy.Provider,
 	})
 }
