+++ builtin/providers/dnsmadeeasy/config.go
@@ -1,4 +1,4 @@
-package main
+package dnsmadeeasy
 
 import (
 	"fmt"
