+++ builtin/providers/dnsmadeeasy/provider.go
@@ -1,4 +1,4 @@
-package main
+package dnsmadeeasy
 
 import (
 	"os"
