+++ builtin/providers/dnsmadeeasy/resource_dme_record.go
@@ -1,4 +1,4 @@
-package main
+package dnsmadeeasy
 
 import (
 	"fmt"
